% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_laguerre.R
\name{assoc_laguerre}
\alias{assoc_laguerre}
\title{assoc_laguerre}
\usage{
assoc_laguerre(n, k, x)
}
\arguments{
\item{n}{The degree.}

\item{x}{The point at which to evaluate the function.}

\item{a}{The generalized parameter.}
}
\value{
The Laguerre(n, a) polynomial at point x.
}
\description{
A utils function
}
