% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_legendre.R
\name{assoc_legendre}
\alias{assoc_legendre}
\title{assoc_legendre}
\usage{
assoc_legendre(l, m, x)
}
\arguments{
\item{l}{The order l.}

\item{m}{The order m.}

\item{x}{The value at which to evaluate the function.}
}
\value{
The P^m_l(x) value.
}
\description{
Computes the value of the associated Legendre polynomial
P(`l`, `m`) at point `x`.
}
\section{special cases}{

When |m| > l, the returned value is 0.
}

