% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binomial.R
\name{binomial}
\alias{binomial}
\title{binomial}
\usage{
binomial(a, k)
}
\arguments{
\item{a}{The first (top) parameter.}

\item{k}{The second (bottom) parameter.}
}
\value{
`a` choose `k`.
}
\description{
A generalized (complex-valued) binomial function.
}
