% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coords.R
\name{car2sph}
\alias{car2sph}
\title{car2sph}
\usage{
car2sph(x, y, z)
}
\arguments{
\item{x}{The cartesian x coordinate.}

\item{y}{The cartesian y coordinate.}

\item{z}{The cartesian z coordinate.}
}
\value{
A vector of length 3, corresponding to { r, theta, phi }.
}
\description{
Transforms cartesian coordinates to spherical.
}
