% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coords.R
\name{complex2phase}
\alias{complex2phase}
\title{complex2phase}
\usage{
complex2phase(c)
}
\arguments{
\item{c}{A complex number.}
}
\value{
A list of `r` and `theta`.
}
\description{
Given a complex number `c`, compute the transformation from
`c = x + iy` to `z = r * exp(i * theta)`.
}
