% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rptSphericalHarmonics.R
\name{rptSphericalHarmonics}
\alias{rptSphericalHarmonics}
\title{rptSphericalHarmonics}
\usage{
rptSphericalHarmonics(l = 1, m = 0, density = 1)
}
\arguments{
\item{l}{The order `l`.}

\item{m}{The degree `m`.}

\item{density}{The density of points being returned. `1` is 100% (default).}
}
\value{
A complex matrix of [ r phi theta ] coordinates.
}
\description{
Computes the Spherical Harmonics for order `l` and degree `m`
and returns a matrix of [ r phi theta ] components.
}
