% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coords.R
\name{sph2car}
\alias{sph2car}
\title{sph2car}
\usage{
sph2car(r, theta, phi)
}
\arguments{
\item{r}{The radius.}

\item{theta}{The azimuthal angle.}

\item{phi}{The polar angle.}
}
\value{
A vector of length 3, corresponding to { x, y, z }.
}
\description{
Transforms spherical coordinates to cartesian.
}
