% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sphericalHarmonicY.R
\name{sphericalHarmonicK}
\alias{sphericalHarmonicK}
\title{sphericalHarmonicK}
\usage{
sphericalHarmonicK(l, m)
}
\arguments{
\item{l}{The degree of the Legendre polynomial.}

\item{m}{The order of the Legendre polynomial.}
}
\value{
The renormalization constant for the Re(SphHm).
}
\description{
Computes the K/N coefficient (renormalization constant) for the Real
Spherical Harmonics.
}
