% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sphericalHarmonicY.R
\name{sphericalHarmonicY}
\alias{sphericalHarmonicY}
\title{sphericalHarmonicY}
\usage{
sphericalHarmonicY(l, m, theta, phi)
}
\arguments{
\item{l}{The degree of the Legendre polynomial.}

\item{m}{The order of the Legendre polynomial.}

\item{theta}{The angle theta at which to evaluate the function at.}

\item{phi}{The angle phi at which to evaluate the function at.}
}
\value{
The spherical harmonic Ylm() evaluated at (theta, phi).
}
\description{
Computes the angular component of the solution.
}
