% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sphericalHarmonicY.R
\name{sphericalHarmonicY2}
\alias{sphericalHarmonicY2}
\title{sphericalHarmonicY2}
\usage{
sphericalHarmonicY2(l, m, theta, phi)
}
\arguments{
\item{l}{The degree of the Legendre polynomial.}

\item{m}{The order of the Legendre polynomial.}

\item{theta}{The angle theta at which to evaluate the function at.}

\item{phi}{The angle phi at which to evaluate the function at.}
}
\value{
The spherical harmonic Y(l, m) evaluated at (theta, phi).
}
\description{
Computes the spherical harmonic of degree l, order m at the
point `P = < theta, phi >`.
}
