% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xyzSphericalHarmonics.R
\name{xyzSphericalHarmonics}
\alias{xyzSphericalHarmonics}
\title{xyzSphericalHarmonics}
\usage{
xyzSphericalHarmonics(l = 1, m = 0, density = 1)
}
\arguments{
\item{l}{The order `l`.}

\item{m}{The degree `m`.}

\item{density}{The density of points being returned. `1` is 100% (default).}
}
\value{
A complex matrix of [ x y z ] coordinates.
}
\description{
Computes the Spherical Harmonics for order `l` and degree `m`
and returns a matrix of [ x y z ] components.
}
