% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sphericalHarmonicK.R
\name{sphericalHarmonicK}
\alias{sphericalHarmonicK}
\title{sphericalHarmonicK}
\usage{
sphericalHarmonicK(l, m)
}
\arguments{
\item{l}{The degree of the Legendre polynomial.}

\item{m}{The order of the Legendre polynomial.}
}
\value{
The re-normalization constant for Y(l, m).
}
\description{
Computes the K coefficient (re-normalization constant) for the Spherical
Harmonics.
}
